requirements.txt
cffi==1.17.1
lxml==4.9.3
numpy==2.1.3
Pillow==10.1.0
pycparser==2.22
PySide6_Essentials==6.8.0
shiboken6==6.8.0
sounddevice==0.5.1
pyqtgraph==0.13.7
websockets==14.1